+++ src/main/java/com/moakiee/thunderbolt/ae2/batch/BatchCopyLimitPattern.java
@@ -2,5 +2,5 @@ package com.moakiee.thunderbolt.ae2.batch;
 
 /** Optional per-pattern upper bound for one batch-provider dispatch. */
 public interface BatchCopyLimitPattern {
-    int maxBatchCopies();
+    long maxBatchCopies();
 }
