+++ src/main/java/com/moakiee/thunderbolt/core/craft/WheelCell.java
@@ -5,5 +5,5 @@ import it.unimi.dsi.fastutil.objects.Object2LongOpenHashMap;
 
 final class WheelCell {
     final Object2LongOpenHashMap<AEKey> outputs = new Object2LongOpenHashMap<>();
-    int copies;
+    long copies;
 }
